+++ src/widgets/layout/ui/Footer.tsx
